requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.1
pandas==2.2.3
praw==7.8.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
xmltodict==0.14.2
